Simply use unified memory instead of fake memory allocator
